requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.7.0
click==8.1.7
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
